Add Buildozer Binaries, 0.6.0.dce8b3c287652cbcaf43c8dd076b3f48c92ab44c (#49)
This adds new buildozer binaries for Linux and Mac. The binaries contain the newly merged
'UseImplicitName' functionality that will allow for a smoother integration of buildozer
into Pants.

New functionality: https://github.com/bazelbuild/buildtools/pull/154

The current latest release of buildozer is 0.6.0. Right now, this does not contain the new functionality. As a solution, the binaries were created locally on Mac and Linux systems to include the new functionality, and given the version: `0.6.0. dce8b3c287652cbcaf43c8dd076b3f48c92ab44c`.

This will be paired with a PR in [pants](https://github.com/pantsbuild/pants/) that changes the reference to the buildozer binaries of this version.