Add cloc 1.80 (#94)
The current version of cloc (version 1.82) has a bug where the --ignored
flag is not working properly ( https://github.com/AlDanial/cloc/issues/401 ).
The pants test for the cloc goal tests this behavior, so as long as this
bug is not fixed upstream, pants will fail this test. I've confirmed
that version 1.80 of cloc does not have this bug, so we should be able
to use it until upstream cloc fixes the bug in a new verson.